requirements.txt
mlflow>=2.11.0
dvc>=3.0.0
scikit-learn>=1.4.0
tensorflow>=2.15.0
opencv-python>=4.8.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=10.0.0
albumentations>=1.3.0
seaborn>=0.12.0
joblib>=1.3.0
pyyaml>=6.0